:truck: Move assignment component into folder.
