+++ src/components/Assignment/AssignmentDetail.vue
