+++ src/views/Assignment/Assignments.vue
@@ -78,37 +78,8 @@
               ></Create>
             </md-tab>
 
-            <md-tab md-label="Delete" md-disabled>
-              <p>
-                Lorem ipsum dolor sit amet consectetur, adipisicing elit. Ullam
-                mollitia dolorum dolores quae commodi impedit possimus qui,
-                atque at voluptates cupiditate. Neque quae culpa suscipit
-                praesentium inventore ducimus ipsa aut.
-              </p>
-              <p>
-                Lorem ipsum dolor sit amet consectetur, adipisicing elit. Ullam
-                mollitia dolorum dolores quae commodi impedit possimus qui,
-                atque at voluptates cupiditate. Neque quae culpa suscipit
-                praesentium inventore ducimus ipsa aut.
-              </p>
-              <p>
-                Lorem ipsum dolor sit amet consectetur, adipisicing elit. Ullam
-                mollitia dolorum dolores quae commodi impedit possimus qui,
-                atque at voluptates cupiditate. Neque quae culpa suscipit
-                praesentium inventore ducimus ipsa aut.
-              </p>
-              <p>
-                Lorem ipsum dolor sit amet consectetur, adipisicing elit. Ullam
-                mollitia dolorum dolores quae commodi impedit possimus qui,
-                atque at voluptates cupiditate. Neque quae culpa suscipit
-                praesentium inventore ducimus ipsa aut.
-              </p>
-              <p>
-                Lorem ipsum dolor sit amet consectetur, adipisicing elit. Ullam
-                mollitia dolorum dolores quae commodi impedit possimus qui,
-                atque at voluptates cupiditate. Neque quae culpa suscipit
-                praesentium inventore ducimus ipsa aut.
-              </p>
+            <md-tab md-label="Delete">
+              <notification-setting></notification-setting>
             </md-tab>
           </md-tabs>
         </md-card-content>
@@ -120,16 +91,17 @@
 <script lang="ts">
 import { Component, Vue, Ref } from "vue-property-decorator";
 import { AssignmentApi } from "@/types/GarnBarnApi/AssignmentApi";
-import DialogBox from "@/components/DialogBox/DialogBox";
 import Layout from "@/layouts/Main.vue";
+import DialogBox from "@/components/DialogBox/DialogBox";
+import NotificationSetting from "@/components/NotificationSetting.vue"
 import DialogBoxComponent from "@/components/DialogBox/DialogBoxComponent.vue";
 import Create from "@/components/Create.vue";
+import UserProfileIcon from "@/components/UserProfileIcon.vue";
+import TagBoxChip from "@/components/Tag/TagBoxChip.vue";
 import GarnBarnApi from "@/services/GarnBarnApi/GarnBarnApi";
 import firebase from "firebase/app";
 import { Assignment } from "@/types/garnbarn/Assignment";
 import { GetAllAssignmentApiNextFunctionWrapper } from "@/types/GarnBarnApi/GarnBarnApiResponse";
-import UserProfileIcon from "@/components/UserProfileIcon.vue";
-import TagBoxChip from "@/components/Tag/TagBoxChip.vue";
 
 @Component({
   components: {
@@ -138,6 +110,7 @@ import TagBoxChip from "@/components/Tag/TagBoxChip.vue";
     Create,
     UserProfileIcon,
     TagBoxChip,
+    NotificationSetting,
   },
 })
 export default class Assignments extends Vue {
